Release v1.20.1

+++ src/utils/temp_folder.ts
@@ -73,5 +73,7 @@ export function getTempFolder(): string {
 
 export function cleanUpTempFolder(): void {
 	const tempFolderPath = platformTempFolder();
-	fs.rmdirSync(tempFolderPath, { recursive: true });
+	if (fs.existsSync(tempFolderPath)) {
+		fs.rmdirSync(tempFolderPath, { recursive: true });
+	}
 }
